had to revert
git-svn-id: http://svn.code.sf.net/p/coda-oss/code/trunk@178 1dbd42e5-0d20-4692-8756-183d5fc073b5

+++ modules/c++/math.linear/include/math/linear/VectorN.h
@@ -186,7 +186,7 @@ public:
     _T dot(const VectorN<_ND>& vec) const
     {
         _T acc(0);
-        for (unsigned int i = 0; i < _ND; ++i)
+        for (size_t i = 0; i < _ND; ++i)
         {
             acc += (*this)[i] * vec[i];
         }
